requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
docker==6.1.3
gitpython==3.1.40
watchdog==3.0.0
websockets==12.0
sse-starlette==1.6.5
chromadb==0.4.24
openai==1.30.0
unidiff==0.7.5
python-dotenv==1.0.1
httpx==0.27.0

# Dev / test
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
